refactor: move security event recorder to internal/db
Addresses the last open follow-up from the PR #357 review. The recorder
lives at the domain layer now so non-HTTP callers — background jobs,
WebSocket auth handlers, future netplay session auth — can emit events
without pulling in gin. The gin-aware adapter in internal/api shrinks to
a ~20-line wrapper that extracts the client IP and request path from
gin.Context and delegates to db.RecordSecurityEvent.

Backend
- New server/internal/db/security_event_recorder.go owns the parameter
  bag (SecurityEventInput, exported), the dedup machinery, and the
  persistence + slog mirror. Zero imports from api or gin.
- New server/internal/db/testing.go exports
  ResetSecurityEventDedupForTest following the igdb package convention
  (ForTest suffix, dedicated file).
- Shrunken server/internal/api/security_event_recorder.go keeps only
  recordSecurityEventCtx, the gin adapter.
- All call sites in auth_handler.go and middleware.go updated to use
  db.SecurityEventInput.

Bug fix rolled into the refactor (code-reviewer flagged it):
The dedup key used to fold UserID=nil into the zero value, which meant
an unauthenticated flood and an authenticated user whose ID happened to
be 0 would collapse into one bucket. The key now tracks a hasUserID
bit so nil and 0 are distinct. Regression test added.

Tests
- 6 new unit tests in the db package cover persistence +
  username_lower auto-population, dedup suppression of middleware
  floods, non-dedup of login events, IP distinction, the new
  nil-vs-zero UserID distinction, and silent DB failure handling.
- Uses a narrow newRecorderTestDB helper that only migrates the
  SecurityEvent table, keeping the test file independent of the api
  package's setupTestEnv.
- Comment at the top of the test file documents why t.Parallel() must
  not be used (global dedup singleton).
- Dedup tests removed from security_event_handler_test.go (moved to
  the db package); TestLoginFailureRecordsSecurityEvent stays as an
  end-to-end integration test for the gin adapter.
